hep: handle special collections correctly
Only records with `980__a:HEP` now get `Literature` as a `_collection`,
records not in HEP don't, in order to properly hide the records not
belonging to the main collection. This requires assigning collections
for the `hep` model based on the contents of the records, instead of
doing it in a filter. Fixes #36.

Signed-off-by: Micha Moskovic <michamos@gmail.com>