hep: readd ensure_hep
Commit 0029b23 forgot to consider that records that we're ingesting from
forms don't yet have the `_collections` key set with the proper value,
so we won't get `980__a:HEP` when sending the record to Legacy.

Note that the existing tests ensure that records that belong to a
special collection only don't become part of the HEP collection, while
records that belonged to both continue belonging to both.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>