hep: handle report_numbers with arXiv source
Previously, `report_numbers` with `source=arXiv` were converted in
`hep2marc` to `037` with `9=arXiv`. This was interpreted on legacy as an
arXiv eprint number, as opposed to an ordinary report number with source
arXiv. This introduces a workaround, converting the `source=arXiv` to
`9=arXiv:reportnumber` instead and doing the opposite in the other
direction.

Signed-off-by: Micha Moskovic <michamos@gmail.com>