hep: do the work of invenio-collections in DoJSON
When migrating thousands of records the receiver in invenio-collections
is too slow for practical use. Given that its job is simply to copy
(and normalize) the value of `special_collections` in `_collections`,
we can simply do this work in DoJSON and detach the receiver during
migration (as suggested by @jmartinm).

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>