hep: move incomplete publication_infos
A ``publication_info`` that contains only a ``journal_title`` would
like to represent that we know that this paper was submitted to that
journal, but not yet published.

Unfortunately this is inferred from clues like the formatting of the
PDF, not known, and is not later updated if the paper is rejected or
published.

This makes searching on this field more difficult, as it generates
false positives that have to be filtered out at search time.

Therefore this commit moves these values to ``public_notes``.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>