hep: correctly set citeable for publication_info
`set_citeable` uses the `is_citeable` function from the
`inspire_schemas` literature builder. That function checks for the
existence of some keys in the `publication_info` to determine whether
the record is citeable. However, all keys in the `publication_info` are
always populated, but the value is `None` in case no value is provided.
This makes `is_citeable` always return true if there is any
`publication_info`.

The solution is to call it after `clean_record` that
strips off keys with `None` values.

Signed-off-by: Micha Moskovic <michamos@gmail.com>